--- .envExample
@@ -1,6 +1,0 @@
-NEXT_PUBLIC_FIREBASE_API_KEY = your api key
-NEXT_PUBLIC_FIREBASE_AUTH_DOMAIN = you auth domain
-NEXT_PUBLIC_FIREBASE_PROJECT_ID = your project id
-NEXT_PUBLIC_FIREBASE_STORAGE_BUCKET = your storage bucket
-NEXT_PUBLIC_FIREBASE_MESSAGING_SENDER_ID = your messaging sender id
-NEXT_PUBLIC_FIREBASE_APP_ID = your app id
