+++ src/utils/base64.ts
@@ -1,3 +1,7 @@
 export const encodeUrlToBase64 = (url: string) => {
   return btoa(encodeURIComponent(url));
 };
+
+export const decodeUrlFromBase64 = (url: string) => {
+  return decodeURIComponent(atob(url));
+};
