--- src/utils/fromBase64.ts
@@ -1,3 +1,0 @@
-export const decodeUrlFromBase64 = (url: string) => {
-  return decodeURIComponent(atob(url));
-};
