+++ src/utils/getDocumentation.ts
@@ -1,5 +1,15 @@
-import { getTypes } from '@/utils/getTypesSchema';
 import { Schema } from '@/types/graphQLSchema';
+import { Documentation } from '@/types/graphQLSchema';
+
+export const getTypes = (documentation: Documentation) => {
+  const types = documentation.data.data.__schema.types
+    .filter((type) => type.kind === 'OBJECT')
+    .map((type) => ({
+      name: type.name,
+      fields: type.fields,
+    }));
+  return types;
+};
 
 export const handleGetDocumentation = async (
   url: string,
